+++ applications/kocolor/features/analyzer/src/main/java/com/zoewave/probase/kocolor/features/analyzer/simulator/ui/StyleSimulatorScreen.kt
@@ -34,9 +34,9 @@ import androidx.compose.ui.unit.sp
 import com.zoewave.probase.core.model.ritual.ClothingCategory
 import com.zoewave.probase.core.model.ritual.ClothingItem
 import com.zoewave.probase.kocolor.features.analyzer.R
-import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.AnalysisStep
-import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.MagicBackground
-import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.MessagingStep
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.graphics.AnalysisStep
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.graphics.MagicBackground
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.list.MessagingStep
 import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.ResultStep
 import com.zoewave.probase.kocolor.model.KoColorRoute
