+++ applications/kocolor/features/analyzer/src/main/java/com/zoewave/probase/kocolor/features/analyzer/simulator/ui/components/graphics/BlueprintCallout.kt
@@ -1,0 +1,129 @@
+package com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.graphics
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.zoewave.probase.features.graphics.colorpicker.util.parseColor
+
+@Composable
+fun BlueprintCallout(
+    label: String,
+    productName: String,
+    colorHex: String?,
+    modifier: Modifier = Modifier
+) {
+    var isExpanded by remember { mutableStateOf(false) }
+
+    Box(modifier = modifier) {
+        Card(
+            modifier = Modifier
+                .width(if (isExpanded) 160.dp else 120.dp)
+                .clickable { isExpanded = !isExpanded },
+            shape = RoundedCornerShape(12.dp),
+            colors = CardDefaults.cardColors(containerColor = Color.White),
+            elevation = CardDefaults.cardElevation(defaultElevation = if (isExpanded) 12.dp else 4.dp)
+        ) {
+            Column(
+                modifier = Modifier.padding(horizontal = 14.dp, vertical = 10.dp),
+                verticalArrangement = Arrangement.spacedBy(4.dp),
+                horizontalAlignment = Alignment.Start
+            ) {
+                Text(
+                    text = label.uppercase(),
+                    style = MaterialTheme.typography.labelSmall.copy(
+                        fontSize = 8.sp,
+                        fontWeight = FontWeight.Bold,
+                        letterSpacing = 2.sp
+                    ),
+                    color = Color.Gray.copy(alpha = 0.6f)
+                )
+
+                if (isExpanded) {
+                    Text(
+                        text = productName,
+                        style = MaterialTheme.typography.titleMedium.copy(
+                            fontSize = 14.sp,
+                            fontFamily = FontFamily.Serif,
+                            lineHeight = 16.sp
+                        ),
+                        color = Color.Black,
+                        maxLines = 2
+                    )
+
+                    // Color Ribbon (All about the color)
+                    if (colorHex != null) {
+                        Spacer(Modifier.height(2.dp))
+                        Box(
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .height(22.dp)
+                                .clip(RoundedCornerShape(4.dp))
+                                .background(
+                                    Brush.horizontalGradient(
+                                        colors = listOf(
+                                            parseColor(colorHex),
+                                            parseColor(colorHex).copy(alpha = 0.7f),
+                                            parseColor(colorHex).copy(alpha = 0.9f)
+                                        )
+                                    )
+                                )
+                                .border(1.dp, Color.Black.copy(alpha = 0.05f), RoundedCornerShape(4.dp))
+                        )
+                    }
+                } else {
+                    Text(
+                        text = "Details",
+                        style = MaterialTheme.typography.labelMedium.copy(fontSize = 11.sp),
+                        color = Color.Black,
+                        textDecoration = androidx.compose.ui.text.style.TextDecoration.Underline
+                    )
+                }
+            }
+        }
+
+        // Color Indicator (The "Anchor Dot")
+        if (colorHex != null) {
+            Box(
+                modifier = Modifier
+                    .align(Alignment.TopEnd)
+                    .offset(x = 6.dp, y = (-6).dp)
+                    .size(18.dp)
+                    .clip(CircleShape)
+                    .background(parseColor(colorHex))
+                    .border(2.dp, Color.White, CircleShape)
+                    .shadow(4.dp, CircleShape)
+            )
+        }
+    }
+}
