+++ applications/kocolor/features/analyzer/src/main/java/com/zoewave/probase/kocolor/features/analyzer/simulator/ui/components/list/MessagingStep.kt
@@ -1,0 +1,186 @@
+package com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.list
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.DirectionsWalk
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.AsyncImage
+import com.zoewave.probase.core.model.ritual.*
+import com.zoewave.probase.kocolor.features.analyzer.R
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.SimulatorEvent
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.StyleSimulatorUiState
+import com.zoewave.probase.kocolor.model.KoColorRoute
+
+@Composable
+fun MessagingStep(
+    uiState: StyleSimulatorUiState,
+    onEvent: (SimulatorEvent) -> Unit,
+    navTo: (KoColorRoute) -> Unit
+) {
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(24.dp),
+        contentPadding = PaddingValues(bottom = 32.dp)
+    ) {
+        item {
+            Spacer(Modifier.height(24.dp))
+            Text(
+                text = stringResource(R.string.applications_kocolor_features_analyzer_simulator_intent_title),
+                style = MaterialTheme.typography.displayMedium.copy(fontSize = 44.sp),
+                fontFamily = FontFamily.Serif,
+                fontWeight = FontWeight.Normal,
+                lineHeight = 52.sp,
+                color = Color.Black
+            )
+        }
+
+        // User Portrait Slot
+        item {
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(100.dp)
+                    .shadow(elevation = 8.dp, shape = RoundedCornerShape(24.dp), ambientColor = Color.Black.copy(alpha = 0.1f)),
+                shape = RoundedCornerShape(24.dp),
+                colors = CardDefaults.cardColors(containerColor = Color.White.copy(alpha = 0.95f)),
+                border = androidx.compose.foundation.BorderStroke(1.dp, Color.White.copy(alpha = 0.5f))
+            ) {
+                Row(
+                    modifier = Modifier.padding(12.dp).fillMaxSize(),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Box(
+                        modifier = Modifier
+                            .size(64.dp)
+                            .clip(CircleShape)
+                            .background(Color.White),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        if (uiState.userPortraitUri != null) {
+                            AsyncImage(
+                                model = uiState.userPortraitUri,
+                                contentDescription = null,
+                                modifier = Modifier.fillMaxSize(),
+                                contentScale = ContentScale.Crop
+                            )
+                        } else {
+                            Icon(Icons.Default.Person, null, modifier = Modifier.size(24.dp), tint = Color.LightGray.copy(alpha = 0.5f))
+                        }
+                    }
+                    Spacer(Modifier.width(16.dp))
+                    Column(modifier = Modifier.weight(1f)) {
+                        Text(
+                            text = if (uiState.userPortraitUri != null) "Visual Identity Active" else "No Portrait Detected",
+                            style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.SemiBold),
+                            color = Color.Black.copy(alpha = 0.8f)
+                        )
+                        Text(
+                            text = "Tap to change visual anchor",
+                            style = MaterialTheme.typography.labelSmall,
+                            color = Color.Gray.copy(alpha = 0.7f)
+                        )
+                    }
+                    
+                    Row(horizontalArrangement = Arrangement.spacedBy(4.dp)) {
+                        IconButton(onClick = { onEvent(SimulatorEvent.CapturePortrait) }) {
+                            Icon(Icons.Default.PhotoCamera, null, tint = Color.DarkGray.copy(alpha = 0.6f), modifier = Modifier.size(20.dp))
+                        }
+                        IconButton(onClick = { onEvent(SimulatorEvent.PickPortrait) }) {
+                            Icon(Icons.Default.Image, null, tint = Color.Black.copy(alpha = 0.6f), modifier = Modifier.size(20.dp))
+                        }
+                    }
+                }
+            }
+        }
+
+        // Clothing Anchors
+        item {
+            AnchorSection(
+                title = stringResource(R.string.applications_kocolor_features_analyzer_simulator_clothing_anchors),
+                categories = listOf(
+                    Triple("Top", Icons.Default.Checkroom, ClothingCategory.TOPS),
+                    Triple("Bottom", Icons.Default.Layers, ClothingCategory.BOTTOMS),
+                    Triple("Shoes", Icons.AutoMirrored.Filled.DirectionsWalk, ClothingCategory.SHOES)
+                ),
+                selectedCategory = uiState.selectedClothingCategory,
+                onCategorySelect = { onEvent(SimulatorEvent.SelectClothingCategory(it as ClothingCategory)) },
+                families = uiState.clothingFamilies,
+                anchoredFamily = uiState.anchoredClothingFamilies[uiState.selectedClothingCategory],
+                onToggle = { onEvent(SimulatorEvent.ToggleClothingFamily(uiState.selectedClothingCategory, it)) },
+                emptyMessage = "No clothes in this category. Tap + in Collection to add pieces."
+            )
+        }
+
+        // Makeup Anchors
+        item {
+            AnchorSection(
+                title = stringResource(R.string.applications_kocolor_features_analyzer_simulator_makeup_anchors),
+                categories = listOf(
+                    Triple("Eyes", Icons.Default.Visibility, MacroCategory.EYES),
+                    Triple("Cheeks", Icons.Default.FaceRetouchingNatural, MacroCategory.DIMENSION),
+                    Triple("Lips", Icons.Default.Face, MacroCategory.LIPS),
+                    Triple("Nails", Icons.Default.PanTool, MacroCategory.NAILS)
+                ),
+                selectedCategory = uiState.selectedCosmeticCategory,
+                onCategorySelect = { onEvent(SimulatorEvent.SelectCosmeticCategory(it as MacroCategory)) },
+                families = uiState.cosmeticFamilies,
+                anchoredFamily = uiState.anchoredCosmeticFamilies[uiState.selectedCosmeticCategory],
+                onToggle = { onEvent(SimulatorEvent.ToggleCosmeticFamily(uiState.selectedCosmeticCategory, it)) },
+                emptyMessage = "No makeup in this category. Tap + in Collection to add products."
+            )
+        }
+        
+        item {
+            Card(
+                modifier = Modifier.fillMaxWidth(),
+                shape = RoundedCornerShape(24.dp),
+                colors = CardDefaults.cardColors(containerColor = Color.White.copy(alpha = 0.85f)),
+                border = androidx.compose.foundation.BorderStroke(1.dp, Color.White.copy(alpha = 0.5f))
+            ) {
+                OutlinedTextField(
+                    value = uiState.userMessage,
+                    onValueChange = { onEvent(SimulatorEvent.UpdateMessage(it)) },
+                    placeholder = { Text(stringResource(R.string.applications_kocolor_features_analyzer_simulator_intent_placeholder), style = MaterialTheme.typography.bodyLarge, modifier = Modifier.alpha(0.5f)) },
+                    modifier = Modifier.fillMaxWidth().height(140.dp),
+                    shape = RoundedCornerShape(24.dp),
+                    colors = OutlinedTextFieldDefaults.colors(
+                        focusedContainerColor = Color.Transparent,
+                        unfocusedContainerColor = Color.Transparent,
+                        focusedBorderColor = Color.Transparent,
+                        unfocusedBorderColor = Color.Transparent
+                    )
+                )
+            }
+        }
+
+        item {
+            Button(
+                onClick = { onEvent(SimulatorEvent.StartSimulation) },
+                modifier = Modifier.fillMaxWidth().height(64.dp),
+                shape = RoundedCornerShape(32.dp),
+                colors = ButtonDefaults.buttonColors(containerColor = Color.Black),
+                elevation = ButtonDefaults.buttonElevation(defaultElevation = 2.dp)
+            ) {
+                Text(stringResource(R.string.applications_kocolor_features_analyzer_simulator_begin_action), style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.Bold, color = Color.White)
+            }
+        }
+    }
+}
